requirements.txt
Django==3.2.9
Faker==11.3.0
Pillow==8.4.0
PyYAML==6.0
boto3==1.20.51
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-q==1.3.9
django-storages==1.12.3
djangorestframework-camel-case==1.3.0
djangorestframework==3.12.4
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.2
uritemplate==4.1.1
